afficher les publications pour chaque utilisateur connecté en tant que presse

+++ src/main/java/tn/esprit/spring/services/implementations/PublicationService.java
@@ -102,7 +102,7 @@ public class PublicationService implements PublicationInterface {
     }
 
     @Override
-    public List<Publication> getPublicationsByUserId(Long userId) {
+    public List<Publication> getPublicationsByUserId(int userId) {
         return publicationRepository.findByUserId(userId);
     }
